+++ modules/backend/Module.php
@@ -1,0 +1,15 @@
+<?php
+
+namespace app\modules\backend;
+
+class Module extends \yii\base\Module
+{
+    public $controllerNamespace = 'app\modules\backend\controllers';
+
+    public function init()
+    {
+        parent::init();
+
+        // custom initialization code goes here
+    }
+}
