+++ modules/backend/controllers/BaseController.php
@@ -1,0 +1,9 @@
+<?php
+namespace app\modules\backend\controllers;
+
+use yii\web\Controller;
+
+class BaseController extends Controller
+{
+    
+}
