+++ modules/backend/controllers/UserController.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace app\modules\backend\controllers;
+
+use app\modules\backend\models\User;
+
+class UserController extends BaseController
+{
+    public function actionUser()
+    {
+        return $this->render('user');
+    }
+
+}
