+++ modules/backend/models/Comment.php
@@ -1,0 +1,9 @@
+<?php
+namespace app\modules\backend\models;
+
+use app\models\Comment as CommentModel;
+
+class Comment extends CommentModel
+{
+
+}
