+++ modules/backend/models/Post.php
@@ -1,0 +1,9 @@
+<?php
+namespace app\modules\backend\models;
+
+use app\models\Post as PostModel;
+
+class Post extends PostModel
+{
+
+}
