+++ modules/backend/models/Tag.php
@@ -1,0 +1,9 @@
+<?php
+namespace app\modules\backend\models;
+
+use app\models\Tag as TagModel;
+
+class Tag extends TagModel
+{
+
+}
