+++ modules/backend/models/User.php
@@ -1,0 +1,9 @@
+<?php
+namespace app\modules\backend\models;
+
+use app\models\User as UserModel;
+
+class User extends UserModel
+{
+
+}
