+++ modules/backend/views/post/index.php
@@ -1,0 +1,7 @@
+<?php
+/**
+ * Created by PhpStorm.
+ * User: leo
+ * Date: 16/5/9
+ * Time: 下午5:30
+ */
